Migrate from serverspec to busser
* (+) Migration guide: https://www.inspec.io/docs/reference/migration/

* InSpec docs: https://www.inspec.io/docs/reference/resources/
* PATH behavior: https://github.com/inspec/inspec/issues/450#issuecomment-317076814

Previously used busser (default verifier):

* Used [busser](https://github.com/test-kitchen/busser) and the [busser-serverspec](https://github.com/test-kitchen/busser-serverspec) plugin.
