+++ cookbooks/cwb-server/test/integration/default/inspec/default_spec.rb
@@ -1,0 +1,5 @@
+title 'cwb-server::default'
+
+describe command('git --version') do
+  its(:stdout) { should match(/git version 2\.\d+\.\d+/) }
+end
