+++ cookbooks/cwb-server/test/integration/default/inspec/nodejs_spec.rb
@@ -1,0 +1,5 @@
+title 'cwb-server::nodejs'
+
+describe command('node --version') do
+  its(:stdout) { should match(/v10\.\d+\.\d+/) }
+end
