+++ cookbooks/cwb-server/test/integration/default/inspec/users_spec.rb
@@ -1,0 +1,12 @@
+title 'cwb-server::users'
+
+describe user('apps') do
+  it { should exist }
+  its('groups') { should eq %w(apps) }
+  it { should have_home_directory '/home/apps' } # Required for Vagrant
+end
+
+describe user('deploy') do
+  it { should exist }
+  its('groups') { should eq %w(deploy apps) }
+end
