--- cookbooks/cwb-server/test/integration/default/serverspec/default_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe 'cwb-server::default' do
-  describe command('git --version') do
-    its(:stdout) { should match(/git version 2\.\d+\.\d+/) }
-  end
-end
