--- cookbooks/cwb-server/test/integration/default/serverspec/nodejs_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe 'cwb-server::nodejs' do
-  describe command('node --version') do
-    its(:stdout) { should match(/v10\.\d+\.\d+/) }
-  end
-end
